+++ src/main/java/chess/console/controller/Command.java
@@ -1,8 +1,8 @@
-package chess.controller;
+package chess.console.controller;
 
 import chess.domain.ChessGame;
 import chess.domain.Color;
-import chess.view.OutputView;
+import chess.console.view.OutputView;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.BiConsumer;
